# src/modules/notification/dtos/notification.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty, IsString } from 'class-validator';

export class NotificationDTO {
  @ApiProperty()
  @IsString()
  title: string;

  @ApiProperty()
  @IsString()
  ledId: string;

  @ApiProperty()
  @IsString()
  event: string;
}