# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "AppController_resendOTP",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/group/create": {
      "post": {
        "operationId": "GroupController_creatGroup",
        "summary": "Create new group records",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create successed!"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "group"
        ]
      }
    },
    "/group/list-group": {
      "get": {
        "operationId": "GroupController_getListGroup",
        "summary": "Get list group records",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get list group successed!"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "group"
        ]
      }
    },
    "/group/detail-group": {
      "post": {
        "operationId": "GroupController_getDetailGroup",
        "summary": "Get detail group records",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetDetailGroupDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Get detail group successed!"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "group"
        ]
      }
    },
    "/group/delete-group": {
      "delete": {
        "operationId": "GroupController_deleteGroup",
        "summary": "Delete group",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteGroupDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Delete group successed!"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "group"
        ]
      }
    },
    "/group/create-led": {
      "post": {
        "operationId": "GroupController_createLed",
        "summary": "Create new led to group records",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLedGroupDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create new led successed!"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "group"
        ]
      }
    },
    "/group/create-schedule": {
      "post": {
        "operationId": "GroupController_createSchedule",
        "summary": "Create new led to group records",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupScheduleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create new schedule successed!"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "group"
        ]
      }
    },
    "/group/update-status": {
      "post": {
        "operationId": "GroupController_updateStatus",
        "summary": "Update group status",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update group status successed!"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "group"
        ]
      }
    },
    "/group/schedule/{id}": {
      "get": {
        "operationId": "GroupController_getScheduleList",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "group"
        ]
      }
    },
    "/group/schedule": {
      "patch": {
        "operationId": "GroupController_updateSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupUpdateScheduleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "group"
        ]
      },
      "delete": {
        "operationId": "GroupController_deleteSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupDeleteScheduleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "group"
        ]
      }
    },
    "/auth/token": {
      "post": {
        "operationId": "AuthController_authToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "Dashboard login",
          "Auth token"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successed!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "Dashboard login"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Register successed!"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Invalid username or password"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "Dashboard login"
        ]
      }
    },
    "/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "summary": "Get user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDetailDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Get successed!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/update-fcm-token/{id}": {
      "patch": {
        "operationId": "UsersController_updateFcmToken",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FcmTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/led/data": {
      "post": {
        "operationId": "LedController_createLed",
        "summary": "Create new led records",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLedDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create successed!"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "led control"
        ]
      },
      "get": {
        "operationId": "LedController_getListData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "led control"
        ]
      }
    },
    "/led/led-detail": {
      "post": {
        "operationId": "LedController_getDetailLed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "led control"
        ]
      }
    },
    "/led/schedule": {
      "post": {
        "operationId": "LedController_createSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "led control"
        ]
      },
      "patch": {
        "operationId": "LedController_updateSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "led control"
        ]
      },
      "delete": {
        "operationId": "LedController_deleteSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteScheduleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "led control"
        ]
      }
    },
    "/led/schedule/{id}": {
      "get": {
        "operationId": "LedController_getScheduleList",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "led control"
        ]
      }
    },
    "/led/brightness": {
      "patch": {
        "operationId": "LedController_updateBrightness",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLumiDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "led control"
        ]
      }
    },
    "/led/update-mode": {
      "patch": {
        "operationId": "LedController_updateMode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedUpdateModeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "led control"
        ]
      }
    },
    "/notifications": {
      "post": {
        "operationId": "NotificationsController_sendNotification",
        "summary": "Test notification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created notification!"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "Notification control"
        ]
      },
      "get": {
        "operationId": "NotificationsController_getNotification",
        "summary": "get notification",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get notification!"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "Notification control"
        ]
      }
    }
  },
  "info": {
    "title": "Led Controller",
    "description": "The led control system",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateGroupDTO": {
        "type": "object",
        "properties": {
          "nameGroup": {
            "type": "string"
          }
        },
        "required": [
          "nameGroup"
        ]
      },
      "GetDetailGroupDTO": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          }
        },
        "required": [
          "groupId"
        ]
      },
      "DeleteGroupDTO": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          }
        },
        "required": [
          "groupId"
        ]
      },
      "CreateLedGroupDTO": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          },
          "lat": {
            "type": "number",
            "example": 10.876143961570557
          },
          "lon": {
            "type": "number",
            "example": 106.80443585691286
          },
          "name": {
            "type": "string",
            "example": "LED_DEMO"
          }
        },
        "required": [
          "groupId",
          "lat",
          "lon",
          "name"
        ]
      },
      "CreateGroupScheduleDTO": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string",
            "example": "6624cf2f9c51b2311c8a02ac"
          },
          "time": {
            "type": "string",
            "pattern": "/^([1-9]|0[1-9]|1[0-2]):[0-5][0-9] ([AaPp][Mm])$/",
            "example": "12:00 AM"
          },
          "value": {
            "type": "number",
            "example": 90
          }
        },
        "required": [
          "groupId",
          "time",
          "value"
        ]
      },
      "UpdateStatusDTO": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "groupId",
          "status"
        ]
      },
      "GroupUpdateScheduleDTO": {
        "type": "object",
        "properties": {
          "scheId": {
            "type": "string",
            "example": "66514686f731c9503c7b4fb6"
          },
          "groupId": {
            "type": "string",
            "example": "6645af562c2ff75648cbeaec"
          },
          "value": {
            "type": "number",
            "example": 50
          },
          "time": {
            "format": "date-time",
            "type": "string",
            "pattern": "/^([1-9]|0[1-9]|1[0-2]):[0-5][0-9] ([AaPp][Mm])$/",
            "example": "12:00 AM"
          },
          "status": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "scheId",
          "groupId",
          "value",
          "time",
          "status"
        ]
      },
      "GroupDeleteScheduleDTO": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string",
            "example": "662533343f4e7bc7f001b533"
          },
          "scheduleId": {
            "type": "string",
            "example": "662533573f4e7bc7f001b535"
          }
        },
        "required": [
          "groupId",
          "scheduleId"
        ]
      },
      "LoginDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RegisterDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "username",
          "password"
        ]
      },
      "UserDetailDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username"
        ]
      },
      "FcmTokenDto": {
        "type": "object",
        "properties": {
          "oldToken": {
            "type": "string",
            "example": ""
          },
          "newToken": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "newToken"
        ]
      },
      "CreateLedDTO": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "example": 10.876143961570557
          },
          "lon": {
            "type": "number",
            "example": 106.80443585691286
          },
          "name": {
            "type": "string",
            "example": "LED_DEMO"
          }
        },
        "required": [
          "lat",
          "lon",
          "name"
        ]
      },
      "CreateScheduleDTO": {
        "type": "object",
        "properties": {
          "ledId": {
            "type": "string",
            "example": "6624cf2f9c51b2311c8a02ac"
          },
          "time": {
            "type": "string",
            "pattern": "/^([1-9]|0[1-9]|1[0-2]):[0-5][0-9] ([AaPp][Mm])$/",
            "example": "12:00 AM"
          },
          "value": {
            "type": "number",
            "example": 90
          }
        },
        "required": [
          "ledId",
          "time",
          "value"
        ]
      },
      "UpdateScheduleDTO": {
        "type": "object",
        "properties": {
          "ledId": {
            "type": "string",
            "example": "6624e9206def0ab815d5d3a6"
          },
          "scheduleId": {
            "type": "string",
            "example": "6624e9206def0ab815d5d3a6"
          },
          "value": {
            "type": "number",
            "example": 50
          },
          "time": {
            "format": "date-time",
            "type": "string",
            "pattern": "/^([1-9]|0[1-9]|1[0-2]):[0-5][0-9] ([AaPp][Mm])$/",
            "example": "12:00 AM"
          },
          "status": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "ledId",
          "scheduleId",
          "value",
          "time",
          "status"
        ]
      },
      "DeleteScheduleDTO": {
        "type": "object",
        "properties": {
          "ledId": {
            "type": "string",
            "example": "662533343f4e7bc7f001b533"
          },
          "scheduleId": {
            "type": "string",
            "example": "662533573f4e7bc7f001b535"
          }
        },
        "required": [
          "ledId",
          "scheduleId"
        ]
      },
      "UpdateLumiDTO": {
        "type": "object",
        "properties": {
          "ledId": {
            "type": "string",
            "example": "662533343f4e7bc7f001b533"
          },
          "value": {
            "type": "number",
            "example": 30
          }
        },
        "required": [
          "ledId",
          "value"
        ]
      },
      "LedUpdateModeDTO": {
        "type": "object",
        "properties": {
          "ledId": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "ledId",
          "status"
        ]
      },
      "NotificationDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "ledId": {
            "type": "string"
          },
          "event": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "ledId",
          "event"
        ]
      }
    }
  }
}